Rename file with Swedish passives to passiveVerbs instead of participles
