+++ packages/yoastseo/src/languageProcessing/languages/sv/config/internal/passiveVerbs.js
